+++ src/main/java/de/uni_jena/cs/fusion/abecto/processor/EquivalentValueMappingProcessor.java
@@ -104,7 +104,7 @@ public class EquivalentValueMappingProcessor extends MappingProcessor<Equivalent
 			}
 
 			// add found correspondences
-			addCorrespondence(this.aspect, correspondingResources);
+			addCorrespondence(correspondingResources);
 		}
 
 	}