+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+* do not use `av:relevantResource` anymore
+
 ## [2.0.1] - 2023-04-15
 
 ### Fixed